Update main.go
PLEASE NOTE: This Code is VERY messy, I don't refactor it because this
release will be only to "kill" the time until the Pull Request
(https://github.com/helm/helm/pull/7021) is merged/ accepted.
But this Code works, and please also note that I had to copy Code from
the official Helm Code-Base to make this working.